白検知継続時間通知クラスの作成途中

+++ TJKun2017/src/jp/co/tdc/epbu/tjkun/measure/VibrationDetection.java
@@ -4,9 +4,9 @@ import jp.co.tdc.epbu.tjkun.device.DeviceFactory;
 
 public class VibrationDetection implements Detection{
 
-	int vibrationDetection = 0;
+	float vibrationDetection = 0;
 
-	public VibrationDetection(int vibration) {
+	public VibrationDetection(float vibration) {
 		this.vibrationDetection = vibration;
 	}
